@@ -9,15 +9,15 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-router-dom": "5.2.0"
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
+    "react-router-dom": "6.0.0"
   },
   "devDependencies": {
     "@types/node": "17.0.31",
-    "@types/react": "18.0.8",
-    "@types/react-dom": "18.0.3",
-    "@vitejs/plugin-react": "1.0.7",
+    "@types/react": "18.0.0",
+    "@types/react-dom": "18.0.0",
+    "@vitejs/plugin-react": "1.3.0",
     "eslint": "8.14.0",
     "eslint-config-prettier": "8.5.0",
     "eslint-config-standard": "17.0.0",
@@ -30,7 +30,7 @@
     "husky": "7.0.4",
     "lint-staged": "12.4.1",
     "prettier": "2.6.2",
-    "vite": "2.8.0"
+    "vite": "2.9.5"
   },
   "lint-staged": {
     "*.{js,jsx}": [
